Lookup server matches email as well as userid
When there are several users accounts that share the same email address
AND the userid for one of those accounts is the same as the email
address, the lookup server will pick a user at random from those
accounts that share email address, causing the user to be logged in to
a random server.

This patch will make it so that usernames are only matched against the
userid, and not the email field.

Fixes: #75
Signed-off-by: Micke Nordin <kano@sunet.se>